awesome-ai-builds/beginner/pdf-qa-chatbot/requirements.txt
faiss-cpu==1.9.0
gradio==5.9.1
langchain-text-splitters==0.3.2
numpy==2.1.3
openai==1.58.1
pypdf==5.1.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pytest==8.3.4
